Delete timeline vuex and use workspace vuex model
