+++ vue/src/store/index.js
@@ -2,9 +2,8 @@ import Vuex from "vuex";
 import Vue from 'vue';
 import setup from './modules/setup';
 import user from './modules/user';
-import timeline from './modules/timeline';
 import employee from './modules/employee'
-import workspace from './modules/workspaces/workspace';
+import workspace from './modules/workspaces';
 import article from './modules/workspaces/article';
 import httpService from "../core/services/httpService";
 import {SHARE_ARTICLE, SHARE_FILE} from "@/core/services/event-bus";
@@ -71,6 +70,5 @@ export default new Vuex.Store({
     employee,
     workspace,
     article,
-    timeline
   }
 });
