--- vue/src/store/modules/timeline/index.js
@@ -1,10 +1,0 @@
-import * as actions from './actions';
-import mutations from './mutations';
-import state from './state';
-
-export default {
-  namespaced: true,
-  actions,
-  mutations,
-  state,
-}
