--- vue/src/store/modules/timeline/mutation-types.js
@@ -1,11 +1,0 @@
-export const CHANGE_LOADING = 'timeline/CHANGE_LOADING';
-export const SHOW_TIMELINE_MODAL = 'timeline/SHOW_TIMELINE_MODAL';
-export const HIDE_TIMELINE_MODAL = 'timeline/HIDE_TIMELINE_MODAL';
-export const GET_TIMELINE_DATA = 'timeline/GET_TIMELINE_DATA';
-export const DELETED_TIMELINE_POST = 'timeline/DELETED_TIMELINE_POST';
-export const ADD_TIMELINE_COMMENT = 'timeline/ADD_TIMELINE_COMMENT';
-export const ADD_TIMELINE_CHILD_COMMENT = 'timeline/ADD_TIMELINE_CHILD_COMMENT';
-export const DELETE_TIMELINE_COMMENT = 'timeline/DELETE_TIMELINE_COMMENT';
-export const DELETE_TIMELINE_CHILD_COMMENT = 'timeline/DELETE_TIMELINE_CHILD_COMMENT';
-export const TIMELINE_LIKE = 'timeline/TIMELINE_LIKE';
-export const TIMELINE_UNLIKE = 'timeline/TIMELINE_UNLIKE';
