--- vue/src/store/modules/timeline/state.js
@@ -1,8 +1,0 @@
-const STATE = {
-  loading: false,
-  showModal: false,
-  timelineData: [],
-  modalTimeline: null,
-}
-
-export default STATE;
