+++ vue/src/store/modules/workspaces/mutation-types.js
@@ -1,0 +1,27 @@
+export const SHOW_WORKSPACE_MODAL = 'workspace/SHOW_WORKSPACE_MODAL';
+export const HIDE_WORKSPACE_MODAL = 'workspace/HIDE_WORKSPACE_MODAL';
+export const GET_WORKSPACES = 'workspace/GET_WORKSPACES';
+export const WORKSPACE_DELETED = 'workspace/WORKSPACE_DELETED';
+export const GET_CURRENT_WORKSPACE = 'workspace/GET_CURRENT_WORKSPACE';
+export const TOGGLE_VIEW_LOADING = 'workspace/TOGGLE_VIEW_LOADING';
+export const GET_ARTICLES = 'workspace/GET_ARTICLES';
+export const TOGGLE_ARTICLES_LOADING = 'workspace/TOGGLE_ARTICLES_LOADING';
+export const SHOW_ARTICLE_MODAL = 'workspace/SHOW_ARTICLE_MODAL';
+export const HIDE_ARTICLE_MODAL = 'workspace/HIDE_ARTICLE_MODAL';
+export const CREATE_ARTICLE = 'workspace/CREATE_ARTICLE';
+export const UPDATE_ARTICLE = 'workspace/UPDATE_ARTICLE';
+export const REMOVE_ARTICLE = 'workspace/REMOVE_ARTICLE';
+
+export const CHANGE_TIMELINE_LOADING = 'workspace/CHANGE_LOADING';
+export const SHOW_TIMELINE_MODAL = 'workspace/SHOW_TIMELINE_MODAL';
+export const HIDE_TIMELINE_MODAL = 'workspace/HIDE_TIMELINE_MODAL';
+export const GET_TIMELINE_DATA = 'workspace/GET_TIMELINE_DATA';
+export const ADD_TIMELINE_POST = 'workspace/ADD_TIMELINE_POST';
+export const UPDATE_TIMELINE_POST = 'workspace/UPDATE_TIMELINE_POST';
+export const DELETED_TIMELINE_POST = 'workspace/DELETED_TIMELINE_POST';
+export const ADD_TIMELINE_COMMENT = 'workspace/ADD_TIMELINE_COMMENT';
+export const ADD_TIMELINE_CHILD_COMMENT = 'workspace/ADD_TIMELINE_CHILD_COMMENT';
+export const DELETE_TIMELINE_COMMENT = 'workspace/DELETE_TIMELINE_COMMENT';
+export const DELETE_TIMELINE_CHILD_COMMENT = 'workspace/DELETE_TIMELINE_CHILD_COMMENT';
+export const TIMELINE_LIKE = 'workspace/TIMELINE_LIKE';
+export const TIMELINE_UNLIKE = 'workspace/TIMELINE_UNLIKE';
