--- vue/src/store/modules/workspaces/workspace/mutation-types.js
@@ -1,13 +1,0 @@
-export const SHOW_WORKSPACE_MODAL = 'workspace/SHOW_WORKSPACE_MODAL';
-export const HIDE_WORKSPACE_MODAL = 'workspace/HIDE_WORKSPACE_MODAL';
-export const GET_WORKSPACES = 'workspace/GET_WORKSPACES';
-export const WORKSPACE_DELETED = 'workspace/WORKSPACE_DELETED';
-export const GET_CURRENT_WORKSPACE = 'workspace/GET_CURRENT_WORKSPACE';
-export const TOGGLE_VIEW_LOADING = 'workspace/TOGGLE_VIEW_LOADING';
-export const GET_ARTICLES = 'workspace/GET_ARTICLES';
-export const TOGGLE_ARTICLES_LOADING = 'workspace/TOGGLE_ARTICLES_LOADING';
-export const SHOW_ARTICLE_MODAL = 'workspace/SHOW_ARTICLE_MODAL';
-export const HIDE_ARTICLE_MODAL = 'workspace/HIDE_ARTICLE_MODAL';
-export const CREATE_ARTICLE = 'workspace/CREATE_ARTICLE';
-export const UPDATE_ARTICLE = 'workspace/UPDATE_ARTICLE';
-export const REMOVE_ARTICLE = 'workspace/REMOVE_ARTICLE';
